#44, #30: Create own Widget for CallTimeConfiguration & code style improvements according lint guidlines
